Fix design-review defects; add reviewed-proposals section
Placement: replace multiply-by-weight HRW scoring with logarithmic
weighted rendezvous (Thaler-Ravishankar). The multiply form is biased —
weights 2:1 gave the heavy disk 75% of keys instead of the proportional
66.7%, compounding with pool size. Implemented FPU-free with an exact
fixed-point Q16 -log2 (square-and-shift), with u128 scores for precision
at the extremes and deterministic tie-breaking independent of candidate
order. New statistical tests pin proportional (1:3 -> 75%) and uniform
(equal weights) selection, with tolerance bands that exclude the old
scoring's output.

ioctl ABI: add explicit _pad to tartine_state on both sides (implicit
padding made sizeof - and thus the _IOR command number - differ between
32- and 64-bit userspace), update TARTINE_IOC_GET_STATE accordingly, add
compat_ptr_ioctl to both fops, and add compile-time layout asserts on
tartine_state and the DiskCandidate FFI struct in both C and Rust so
drift fails the build. Gate TARTINE_IOC_MAKE_WRITABLE behind
inode_owner_or_capable() - conversion changes write semantics for every
open fd, so it's restricted like chmod.

Types: align ChunkPointer/Extent/Record checksums to u32 crc32c,
matching DESIGN.md §5.3.

DESIGN.md: fix defects in place - corrected placement formula plus
full-disk preference-list/ENOSPC/statfs semantics (§7.2), fence
durability ordering before the first degraded write plus refuse-on-
ambiguity mount rule (§6), bounded conversion delta passes with a brief
append quiesce so conversion always terminates (§9.3), ioctl ABI and
permission rules (§9.2). Add §16 with reviewed architectural proposals:
self-describing chunk-log with a rebuildable chunk index (drops the
per-append sync metadata commit), placement-group indirection, page-
cache-backed reads, a crash-consistency contract with dm-log-writes/
xfstests testing, and cheap snapshots for append-only files.

+++ crates/tartine-proto/src/lib.rs
@@ -120,7 +120,9 @@ pub struct ChunkPointer {
     pub chunk_seq: u64,
     pub replicas: Vec<DiskId>,
     pub len: u32,
-    pub checksum: u64,
+    /// crc32c of the payload (DESIGN.md §4.1/§5.3 — the kernel's own
+    /// hardware-accelerated crc32c(), so 32 bits, not a 64-bit hash).
+    pub checksum: u32,
 }
 
 /// One entry in a writable file's extent map: a fixed-size, block-aligned
@@ -130,7 +132,8 @@ pub struct Extent {
     pub file_offset: u64,
     pub len: u32,
     pub replicas: Vec<DiskId>,
-    pub checksum: u64,
+    /// crc32c of the extent contents (same scheme as `ChunkPointer`).
+    pub checksum: u32,
 }
 
 #[derive(Debug, Clone)]
